+++ app/src/main/java/net/squanchy/contest/ContestStandings.java
@@ -5,11 +5,11 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class ContestStandings {
 
-    public static ContestStandings create(int goal, int current) {
+    public static ContestStandings create(long goal, int current) {
         return new AutoValue_ContestStandings(goal, current);
     }
 
-    public abstract int goal();
+    public abstract long goal();
 
     public abstract int current();
 }
